Moved tests for the default provider into its own class
